emoteam-auth/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
Flask-Cors==4.0.0
frozenlist==1.4.1
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
six==1.16.0
spotipy==2.23.0
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.18.1
